Avoid failure if 2 userid-customcertid combos
We somehow have two results for
SELECT * FROM m_customcert_issues WHERE userid = 20426 AND customcertid = 35

   	ID	USERID	CUSTOMCERTID	CODE	EMAILED	TIMECREATED
1	8127	20426	35	1cxhyyRCB6	0	1535058946
2	8128	20426	35	Tg7yyHT9rh	0	1535058945

which is causing a fatal error preventing sending any certificate emails afterward.  Not sure if this is from a restore or the user double-clicking submit somewhere.

For those who have a similar issue, this fix gets around the problem by choosing the entry with the max datestamp.