#513 Fix: problem repositioning elements
This commit adds a workaround to the YUI code to handle the fact that
node.getX()/getY() return coordinates based on the current browser
viewport rather than the page.  This had the result that if the
rearrange.php page has been scrolled isoutofbounds() may return true
-- preventing the element from being repositioned -- even though it is
within the PDF's area.  This fix re-reads the PDF area's position within
isoutofbounds() so that the coordinates for the PDF area have the same
system as the element being moved.  This change is required for Moodle
4.0, in Moodle 3.11 the coordinates system between PDF area and elements
remains consistent despite any scrolling.  I have no idea where this
change in behaviour has originated (YUI lib? CSS? Something else?).