Make index not unique (#537)
The reason for this is because there may be plans
to extend this plugin to allow for multiple
certificate issues.

Also, there was a bug that allowed for duplicate
issues which was resolved but those issues could
still exist in users' databases. Instead of creating
an upgrade path to delete them just leave it for now.